pyproject.toml
[project]
name = "cubiai"
version = "0.1.0"
description = "AI-assisted pipeline for converting illustrations into Live2D assets."
readme = "README.md"
license = { file = "LICENSE" }
authors = [
    { name = "JakkuSakura", email = "qjk2001@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "typer[all]>=0.12.3",
    "rich>=13.7.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.1",
    "pyyaml>=6.0.1",
    "Pillow>=10.3.0",
    "numpy>=1.26.4",
    "scikit-image>=0.22.0",
    "opencv-python>=4.9.0.80",
    "orjson>=3.10.6",
    "psd-tools>=1.9.32",
    "scipy>=1.11.0",
    "httpx>=0.27.0"
]

[project.optional-dependencies]
gui = [
    "PySide6>=6.7.0",
    "qasync>=0.26.0"
]
dev = [
    "pytest>=8.2.0"
]

[project.scripts]
cubiai = "cubiai.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
managed = true

[tool.uv.workspace]
members = []
